feat(frontend): add header/footer shell with SVG logo; rename FileUpload to file-upload
- Introduced **site chrome**:
  - Added `Header` with logo (Next/Image) and brand title.
  - Added `Footer` with current year and GitHub link.
  - Wrapped pages in a centered container with consistent spacing.
- Added **branding assets**:
  - `frontend/public/budget-wise-logo.svg`
  - `frontend/public/budget-wise-logo.min.svg`
- Renamed upload component to **kebab-case** for consistency:
  - `FileUpload.tsx` → `file-upload.tsx`
  - Updated imports in `app/page.tsx`.
- Updated `app/layout.tsx` to render the new Header/Footer and set base layout classes.
- chore(makefile): clarify `run-db` help text to “Start Postgres database (dev)”.
